chore: remove AdultsChanged event and ahndle everything inside routing
