+++ sn/src/node/node_api/split.rs
@@ -20,6 +20,6 @@ impl Node {
             )));
         }
 
-        self.level_up().await
+        Ok(())
     }
 }
