+++ sn/src/routing/core/mod.rs
@@ -26,7 +26,7 @@ mod split_barrier;
 
 pub(crate) use back_pressure::BackPressure;
 pub(crate) use bootstrap::{join_network, JoiningAsRelocated};
-pub(crate) use capacity::{CHUNK_COPY_COUNT, MIN_LEVEL_WHEN_FULL};
+pub(crate) use capacity::MIN_LEVEL_WHEN_FULL;
 pub(crate) use chunk_store::ChunkStore;
 pub(crate) use comm::{Comm, ConnectionEvent, SendStatus};
 pub(crate) use proposal::ProposalUtils;
