+++ sn/src/routing/core/msg_handling/anti_entropy.rs
@@ -64,11 +64,17 @@ impl Core {
             )
             .await?;
 
-        self.fire_node_event_for_any_new_adults().await?;
+        let mut commands = vec![];
+
+        commands.extend(self.fire_node_event_for_any_new_adults().await?);
 
         // always run this, only changes will trigger events
-        self.update_self_for_new_node_state_and_fire_events(snapshot)
-            .await
+        commands.extend(
+            self.update_self_for_new_node_state_and_fire_events(snapshot)
+                .await?,
+        );
+
+        Ok(commands)
     }
 
     pub(crate) async fn handle_anti_entropy_retry_msg(
